refactor: chat 도메인에 FSD 구조 적용 (#69)
