+++ src/entities/chat-room/model/store.ts
