+++ src/entities/message/lib/formatter.ts
@@ -1,0 +1,7 @@
+export const formatTimeString = (date: Date) => {
+  return date.toLocaleTimeString('ko-KR', {
+    hour: 'numeric',
+    minute: '2-digit',
+    hour12: true,
+  })
+}
