--- src/features/chat/index.ts
@@ -1,15 +1,0 @@
-import ChatRoomFilter from '@/features/chat/ui/ChatRoomFilter'
-import ChatRoomList from '@/features/chat/ui/ChatRoomList'
-import ChatHeader from '@/features/chat/ui/ChatHeader'
-import ChatSidebar from '@/features/chat/ui/ChatSidebar'
-import MessageInput from '@/features/chat/ui/MessageInput'
-import MessageList from '@/features/chat/ui/MessageList'
-
-export {
-  ChatRoomFilter,
-  ChatRoomList,
-  ChatHeader,
-  ChatSidebar,
-  MessageInput,
-  MessageList,
-}
