+++ src/shared/ui/EmptyState.tsx
