+++ src/shared/ui/Error.tsx
