+++ src/widgets/chat-room-list/lib/formatter.ts
