+++ src/widgets/chat-sidebar/ui/ChatSidebar.tsx
@@ -3,8 +3,8 @@
 import { cn } from '@/shared/lib/cn'
 import { UserIcon } from 'lucide-react'
 import Image from 'next/image'
-import { useChatRoomList } from '@/features/chat/api/queries'
-import { useChatStore } from '@/features/chat/model/store'
+import { useChatStore } from '@/entities/chat-room/model/store'
+import { useChatRoomList } from '@/entities/chat-room/api/queries'
 
 function ChatSidebar() {
   const currentRoomId = useChatStore((state) => state.enteredRoomId)
